refactor: simplify isRandomMove and getWinner functions for readability
