+++ src/const/errorMessage.js
@@ -2,4 +2,5 @@ export const ERROR_MESSAGE = {
   LENGTH_ERROR: '[ERROR] 길이가 5글자를 초과했습니다',
   LENGTH_CARCOUNT_ERROR: `[ERROR] 자동차 개수 100개를 초과했습니다`,
   COUNT_ERROR: '[ERROR] 횟수가 100000회를 초과했습니다',
+  REPEAT_ERROR: '[ERROR] 버스의 이름이 반복되었습니다',
 };
